Added balance refreshing in BalanceService

+++ TechnicalTask/src/main/java/com/malykhnik/technicaltask/model/BankAccount.java
@@ -2,6 +2,8 @@ package com.malykhnik.technicaltask.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 @Data
 @Entity
@@ -13,6 +15,6 @@ public class BankAccount {
     private Long id;
 
     @Column(nullable = false)
-    private int balance;
+    private double balance;
 
 }
